clean package names.
