+++ lib/msf/core/modules/external/java/src/core/src/main/java/com/metasploit/java_external_module/core/function/ThrowableFunction.java
@@ -1,0 +1,13 @@
+package com.metasploit.java_external_module.core.function;
+
+public interface ThrowableFunction<T, R> {
+
+    /**
+     * Applies this function to the given argument.
+     *
+     * @param t the function argument
+     * @return the function result
+     * @throws Throwable when error
+     */
+    R apply(T t) throws Throwable;
+}
