--- lib/msf/core/modules/external/java/src/core/src/main/java/com/xenoamess/metasploit/java_external_module/core/constants/MsfPluginResourceUrlStrings.java
@@ -1,6 +1,0 @@
-package com.xenoamess.metasploit.java_external_module.core.constants;
-
-public class MsfPluginResourceUrlStrings {
-    public static String PARENT = "msf_plugin_resources";
-    public static String METADATA = PARENT + "/metadata.json";
-}
